srcs/backend/requirements.txt
flask
Flask-CLI
mysql-connector
flask-restplus